Added version with colors
This version uses the "Colorama" module, which requires installation so i uploaded it as a seperate file in case anyone doesn't want to deal with it.

Also note that, due to the nature of the program allowing several words of information, this might not be very useful since some misplaced letters will be (not be) on the same slot, but the visualisation can only pick one, because of that, this is mostly intended for single words of information.